+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog
 =========
 
+* Add Django 4.2+ fixer to rewrite ``index_together`` declarations into ``indexes`` declarations in model ``Meta`` classes.
+
+  `PR #464 <https://github.com/adamchainz/django-upgrade/pull/464>`__.
+
 * Fix tracking of AST node parents.
   This may have fixed some subtle bugs in these fixers: